♻️ refactor: ProductDetailInfo 컴포넌트 코드 구조 개선 - 재사용되는 HTML 요소 통일 - 조건에 따라 변경되는 부분의 HTML 요소만 삼항연산자를 활용하여 수정 - getDiffDays 유틸 함수는 별도 파일로 분리
